AzureFunction/requirements.txt
azure-functions==1.4.0
Pillow==7.2.0
python-barcode==0.13.1
